requirements.txt
torch==2.3.1
torchvision==0.18.1
timm==1.0.7
albumentations==1.4.10
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
matplotlib==3.9.0
seaborn==0.13.2
Pillow==10.3.0
pyyaml==6.0.1
tqdm==4.66.4
tensorboard==2.17.0
onnx==1.16.1
onnxruntime==1.18.1
